+++ packages/iov-bns/types/decode.d.ts
@@ -1,10 +1,10 @@
 import { ChainId } from "@iov/base-types";
-import { Amount, BaseTx, BcpNonce, BcpTicker, SignedTransaction, UnsignedTransaction } from "@iov/bcp-types";
+import { Amount, BaseTx, BcpTicker, Nonce, SignedTransaction, UnsignedTransaction } from "@iov/bcp-types";
 import * as codecImpl from "./generated/codecimpl";
 import { BnsBlockchainNft, BnsUsernameNft, Keyed } from "./types";
 export declare function decodeBlockchainNft(nft: codecImpl.blockchain.IBlockchainToken): BnsBlockchainNft;
 export declare function decodeUsernameNft(nft: codecImpl.username.IUsernameToken): BnsUsernameNft;
-export declare function decodeNonce(acct: codecImpl.sigs.IUserData & Keyed): BcpNonce;
+export declare function decodeNonce(acct: codecImpl.sigs.IUserData & Keyed): Nonce;
 export declare function decodeToken(data: codecImpl.namecoin.IToken & Keyed): BcpTicker;
 export declare function decodeAmount(coin: codecImpl.x.ICoin): Amount;
 export declare function parseTx(tx: codecImpl.app.ITx, chainId: ChainId): SignedTransaction;