+++ packages/iov-core/src/multichainsigner.ts
@@ -78,11 +78,16 @@ export class MultiChainSigner {
     return this.getChain(chainId).codec.keyToAddress(key);
   }
 
-  // getNonce will return one value for the address, 0 if not found
-  // not the ful bcp info.
+  /**
+   * Returns one value for the address, 0 if not found.
+   *
+   * This is done automatically when you use signAndCommit().
+   *
+   * @todo This is not tested. Decide if we need to expose this method.
+   */
   public async getNonce(chainId: ChainId, addr: Address): Promise<Nonce> {
     const nonce = await this.getChain(chainId).connection.getNonce({ address: addr });
-    return nonce.data.length === 0 ? (new Int53(0) as Nonce) : nonce.data[0].nonce;
+    return nonce.data.length === 0 ? (new Int53(0) as Nonce) : nonce.data[0];
   }
 
   // signAndCommit will sign the transaction given the signer specified in
@@ -93,14 +98,13 @@ export class MultiChainSigner {
     const chainId = tx.chainId;
     const { connection, codec } = this.getChain(chainId);
 
-    const signer = tx.signer;
-    const signerAddr = this.keyToAddress(chainId, signer);
-    const nonce = await this.getNonce(chainId, signerAddr);
+    const nonceResponse = await this.getChain(chainId).connection.getNonce({ pubkey: tx.signer });
+    const nonce = nonceResponse.data.length === 0 ? (new Int53(0) as Nonce) : nonceResponse.data[0];
 
     // We have the publickey bundle from the transaction, but need
     // a PublicIdentity to sign. Same information content, so I fake it.
     // TODO: Simon, a cleaner solution would be nicer. How?
-    const fakeId: PublicIdentity = { pubkey: signer };
+    const fakeId: PublicIdentity = { pubkey: tx.signer };
     const signed = await this.profile.signTransaction(walletId, fakeId, tx, codec, nonce);
     const txBytes = codec.bytesToPost(signed);
     const post = await connection.postTx(txBytes);