+++ packages/iov-ethereum/src/ethereumconnection.ts
@@ -1,14 +1,13 @@
 import axios from "axios";
 import { Stream } from "xstream";
 
-import { Algorithm, ChainId, PostableBytes, PublicKeyBytes, TxId } from "@iov/base-types";
+import { ChainId, PostableBytes, TxId } from "@iov/base-types";
 import {
   Address,
   BcpAccount,
   BcpAccountQuery,
   BcpBlockInfo,
   BcpConnection,
-  BcpNonce,
   BcpQueryEnvelope,
   BcpQueryTag,
   BcpTicker,
@@ -18,12 +17,15 @@ import {
   ConfirmedTransaction,
   dummyEnvelope,
   isAddressQuery,
+  isPubkeyQuery,
+  Nonce,
   TokenTicker,
 } from "@iov/bcp-types";
 import { Encoding } from "@iov/encoding";
 import { DefaultValueProducer, ValueAndUpdates } from "@iov/stream";
 
 import { constants } from "./constants";
+import { keyToAddress } from "./derivation";
 import { Parse } from "./parse";
 import { decodeHexQuantity, decodeHexQuantityNonce, decodeHexQuantityString, hexPadToEven } from "./utils";
 
@@ -144,39 +146,26 @@ export class EthereumConnection implements BcpConnection {
     return dummyEnvelope(accounts);
   }
 
-  public async getNonce(query: BcpAccountQuery): Promise<BcpQueryEnvelope<BcpNonce>> {
+  public async getNonce(query: BcpAccountQuery): Promise<BcpQueryEnvelope<Nonce>> {
+    let address: Address;
+
     if (isAddressQuery(query)) {
-      const address = query.address;
-
-      // see https://github.com/ethereum/wiki/wiki/JSON-RPC#eth_gettransactioncount
-      const nonceResponse = await axios.post(this.baseUrl, {
-        jsonrpc: "2.0",
-        method: "eth_getTransactionCount",
-        params: [address, "latest"],
-        id: 4,
-      });
-
-      const nonce: BcpNonce = {
-        address: address,
-        // fake pubkey, we cannot always know this
-        pubkey: {
-          algo: Algorithm.Ed25519,
-          data: new Uint8Array([]) as PublicKeyBytes,
-        },
-        nonce: decodeHexQuantityNonce(nonceResponse.data.result),
-      };
-
-      const out: BcpQueryEnvelope<BcpNonce> = {
-        metadata: {
-          offset: 0,
-          limit: 0,
-        },
-        data: [nonce],
-      };
-      return Promise.resolve(out);
+      address = query.address;
+    } else if (isPubkeyQuery(query)) {
+      address = keyToAddress(query.pubkey);
     } else {
       throw new Error("Query type not supported");
     }
+
+    // see https://github.com/ethereum/wiki/wiki/JSON-RPC#eth_gettransactioncount
+    const nonceResponse = await axios.post(this.baseUrl, {
+      jsonrpc: "2.0",
+      method: "eth_getTransactionCount",
+      params: [address, "latest"],
+      id: 4,
+    });
+
+    return Promise.resolve(dummyEnvelope([decodeHexQuantityNonce(nonceResponse.data.result)]));
   }
 
   public changeBlock(): Stream<number> {
@@ -187,7 +176,7 @@ export class EthereumConnection implements BcpConnection {
     throw new Error("Not implemented");
   }
 
-  public watchNonce(_: BcpAccountQuery): Stream<BcpNonce | undefined> {
+  public watchNonce(_: BcpAccountQuery): Stream<Nonce | undefined> {
     throw new Error("Not implemented");
   }